Added absence as a background service #14
